Initial import of a data model for an LDAP directory to be used with the Interview framework.
There is still much to do, but it's a start. So far the model provides 
data showing the hierarchical structure of your LDAP directory and the
objectClasses of the objects in the directory (tooltips at present).

The model also uses lazy fetching of the data to prevent long startup
times.

The eventual plan is for the modle to hold all data of the directory and 
to filter it by means of proxy models. This will allow us to use the 
same model to for e.g. Show an overview of the directory structure in 
one view, and the details of the selected LDAP object in another view. 
At present the model only searches for the DN and objectClass 
attributes - this will be extended shortly.

Also added the class LdapDN for easy manipulation of distinguished 
names.

svn path=/trunk/KDE/kdepimlibs/; revision=620873