kldap: Fix LDAP spelling to pull in #includes.
The recent CMake improvements changed the spelling of "LDAP" to be
"Ldap", and while this may be desirable CMake style, CMake ends up
maintaining the case when expanding cmakedefines later.

This cases the C++ source to fail the "#ifdef LDAP_FOUND" and similar
and skip #including <ldap.h>, and therefore failing to link.

This could be fixed by modifying all usages of LDAP_FOUND and similar,
I've chosen instead to revert just the affected portion of the
CMakeLists.txt, hope this doesn't bug anyone.